I fixed the requirements

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,12 +5,12 @@
  */
 int main(void)
 {
-	long long int n1 = 0, n2 = 1, count = 0, fib = 1;
+	long unsigned int n1 = 0, n2 = 1, count = 0, fib = 1;
 
 	while (count <= 50)
 	{
 		fib = n1 + n2;
-		printf("%lld", fib);
+		printf("%lu", fib);
 		if (count < 50)
 		{
 			printf(", ");
